added a new files where the evaluation function have been put into
