+++ query_handling.py
@@ -13,42 +13,9 @@ import lda
 import preprocessing
 import utility
 
-cv_matrix = sp.load_npz("Generated Files/count_vec_matrix.npz")
-dt_matrix = sp.load_npz("Generated Files/topic_doc_matrix.npz")
-tw_matrix = sp.load_npz("Generated Files/topic_word_matrix.npz")
 doc2word = utility.load_vector_file("Generated Files/doc2word.csv")
 word2vec = utility.load_vector_file("Generated Files/word2vec.csv")
 inverse_w2v = {v: k for k, v in word2vec.items()}
-dirichlet_smoothing = sum([len(i) for i in list(doc2word.values())]) / len(doc2word)
-
-
-def lda_evaluate_word_doc(document_index, word_index):
-    """
-    The LDA evaluates a document against a word and returns the score
-    :param document_index: document
-    :param word_index: word
-    :return: a score
-    """
-    word_topics = tw_matrix.getcol(word_index)
-    doc_topics = dt_matrix[document_index].T
-    score = word_topics.multiply(doc_topics).sum()
-    return score
-
-
-def lm_evaluate_word_doc(document_index, word_index):
-    """
-    The language model evaluates a document against a word and returns the score
-    :param document_index: document
-    :param word_index: word
-    :return: a score
-    """
-    N_d = len(doc2word[document_index])
-    tf = cv_matrix[document_index, word_index]
-    w_freq_in_D = np.sum(cv_matrix[:, word_index])
-    number_of_word_tokens = len(word2vec)
-    score = np.prod([(N_d / (N_d + dirichlet_smoothing)), (tf / N_d)]) + \
-           np.prod([(1 - (N_d / (N_d + dirichlet_smoothing))), (w_freq_in_D / number_of_word_tokens)])
-    return score
 
 
 def evaluate_query_doc(function, query: List[str], document_index: int):
@@ -78,7 +45,7 @@ def evaluate_query(function, query_index, query_words, tell=False):
     """
     lst = {}
     with Pool(processes=8) as p:
-        max_ = cv_matrix.shape[0]
+        max_ = len(doc2word)
         with tqdm(total=max_) as pbar:
             for i, score in enumerate(
                     p.imap(partial(evaluate_query_doc, function, query_words), range(0, max_))):
